+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [vNext]
 ### Updated
 - Updated DOM Snapshot to 3.6.0 and supported scripts for internet explorer. [Trello 1962](https://trello.com/c/MlHqSdXv)
+- The default scroll root element is now the bigger one between "body" and "html" instead of only "html". [Trello 1972](https://trello.com/c/YfJRReVo)
 
 ## [3.168.1]  - 2020-06-28
 ### Fixed